GitLab scoping search input
